+++ examples/dart_examples/bin/timer.dart
@@ -1,9 +1,11 @@
 import 'package:signals/signals.dart';
 
+/// Trigger [TimerSignalEvent]
 main() {
-  final timer = Duration(milliseconds: 2000).toSignal(cancelOnError: true);
+  final timer = TimerSignal(every: Duration(seconds: 1));
 
   effect(() {
+    // Receive a TimerSignalEvent
     final tick = timer();
     final dateTime = DateTime.fromMillisecondsSinceEpoch(tick.millis);
     print("$tick => ${dateTime.hour}:${dateTime.minute}:${dateTime.second}");